requirements.txt
pydantic
omegaconf
#hydra-core
qdrant-client
psycopg2-binary
duckdb
python-dotenv
fastapi
uvicorn
pandas